pyproject.toml
[tool.poetry]
name = "llm-swarm"
version = "0.1.0"
description = "A swarm of large language models that love to eat"
authors = ["somewheresy <warchest@dataclysm.xyz>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
together = "^1.3.11"
python-dotenv = "^1.0.0"
colorama = "^0.4.6"
chromadb = "^0.4.24"
numpy = "^1.26.0"
sentence-transformers = "^2.5.1"
gradio = "^4.19.2"
pandas = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
llm-swarm = "src.llm_swarm.main:main" 